+++ CHANGELOG.md
@@ -231,12 +231,16 @@ Compare to [stable](https://github.com/JonathanCasey/grand_trade_auto/compare/st
       requirements ([#2][]).
 - [Added] `databases` file started, loads a default main database handle from
       config ([#2][]).
+- [Changed] `DatabaseMeta` is now simply `Database` ([#80][]).
 
 
 ### Database: Postgres
 - [Added] `DatabasePostgres` file and class started, with ability to load from
       config, connect, create/drop/check-exists DB ([#2][]).
 - [Added] Logging setup and logger usage added to existing code ([#8][]).
+- [Changed] `database_postgres` and `test_database_postgres` modules are now
+      simply `postgres` and `test_postgres`, respectively ([#80][]).
+- [Changed] `DatabasePostgres` is now simply `Postgres` ([#80][]).
 
 
 ### Datafeeds / Meta
@@ -440,6 +444,7 @@ Compare to [stable](https://github.com/JonathanCasey/grand_trade_auto/compare/st
 - [#73][]
 - [#75][]
 - [#77][]
+- [#80][]
 
 #### PRs
 - [#29][] for [#26][]
@@ -469,6 +474,7 @@ Compare to [stable](https://github.com/JonathanCasey/grand_trade_auto/compare/st
 - [#76][] for [#74][]
 - [#78][] for [#77][]
 - [#79][] for [#7][], [#54][]
+- [#83][] for [#80][]
 
 
 ---
@@ -505,6 +511,7 @@ Reference-style links here (see below, only in source) in develop-merge order.
 [#77]: https://github.com/JonathanCasey/grand_trade_auto/issues/77 'Issue #77'
 [#7]: https://github.com/JonathanCasey/grand_trade_auto/issues/7 'Issue #7'
 [#54]: https://github.com/JonathanCasey/grand_trade_auto/issues/54 'Issue #54'
+[#80]: https://github.com/JonathanCasey/grand_trade_auto/issues/80 'Issue #80'
 
 [#29]: https://github.com/JonathanCasey/grand_trade_auto/pull/26 'PR #29'
 [#30]: https://github.com/JonathanCasey/grand_trade_auto/pull/30 'PR #30'
@@ -533,3 +540,4 @@ Reference-style links here (see below, only in source) in develop-merge order.
 [#76]: https://github.com/JonathanCasey/grand_trade_auto/pull/76 'PR #76'
 [#78]: https://github.com/JonathanCasey/grand_trade_auto/pull/78 'PR #78'
 [#79]: https://github.com/JonathanCasey/grand_trade_auto/pull/79 'PR #79'
+[#83]: https://github.com/JonathanCasey/grand_trade_auto/pull/83 'PR #83'